requirement.txt
3==0.0.0
alabaster==0.7.12
allure-pytest==2.8.16
allure-python-commons==2.8.16
apipkg==1.5
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
behave==1.2.6
boto3==1.10.41
botocore==1.13.41
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.1.2
colour==0.1.5
config==0.4.2
cryptography==2.7
cycler==0.10.0
dataframe==0.2.1.3
dbapi==0.0.14
ddt==1.2.1
docutils==0.15.2
driver==0.1.0
enum34==1.1.10
et-xmlfile==1.0.1
execnet==1.7.1
html-testRunner==1.2.1
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
JayDeBeApi==1.1.1
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
JPype1==0.7.0
jsonschema==3.0.2
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==7.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.17
namedlist==1.7
nose==1.3.7
numpy==1.17.1
objectpath==0.6.1
openpyxl==2.6.3
packaging==19.1
pandas==0.25.0
paramiko==2.6.0
parse==1.12.1
parse-type==0.5.2
pluggy==0.12.0
protobuf==3.9.1
psycopg2-binary==2.8.3
py==1.8.0
pycparser==2.19
Pygments==2.4.2
pymongo==3.9.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.4.3
pytest-asyncio==0.10.0
pytest-forked==1.1.3
pytest-ordering==0.6
pytest-parallel==0.1.0
pytest-repeat==0.8.0
pytest-rerunfailures==9.0
pytest-xdist==1.31.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
robotframework==3.1.2
robotframework-seleniumlibrary==3.3.1
s3transfer==0.2.1
scipy==1.3.1
selenium==3.141.0
six==1.12.0
snowballstemmer==1.9.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.7
sshtunnel==0.1.5
StyleFrame==2.0.4
tabulate==0.8.3
tblib==1.6.0
tesults==1.2.1
tqdm==4.36.1
urllib3==1.25.3
wcwidth==0.1.7
webdrivermanager==0.7.4
xlrd==1.2.0
XlsxWriter==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==0.5.2
